后端适配、跨帧去重机制修改

+++ backend/src/main/java/org/example/roaddetection/dto/AiDetectionItem.java
@@ -10,6 +10,6 @@ public class AiDetectionItem {
     private Integer class_id;
     private String class_name;
     private Double confidence;
-    private List<Float> bbox;
+    private List<Double> bbox;
     private FeatureDTO feature;
 }
